+++ divio_cli/api_requests.py
@@ -1,8 +1,9 @@
 import os
 
+from six.moves.urllib_parse import urljoin, urlparse
+
 import click
 import requests
-from six.moves.urllib_parse import urljoin, urlparse
 
 from . import messages
 from .utils import create_temp_dir, get_user_agent
@@ -322,7 +323,7 @@ class UploadDBRequest(JsonResponse, APIRequest):
     url = "/api/v1/website/{website_id}/upload/db/"
     method = "POST"
 
-    def get_error_code_map(self):
+    def get_error_code_map(self, login=None):
         error_codes = super(UploadDBRequest, self).get_error_code_map()
         error_codes[requests.codes.bad_request] = messages.INVALID_DB_SUBMITTED
         return error_codes
@@ -368,3 +369,10 @@ class GetCustomEnvironmentVariablesRequest(JsonResponse, APIRequest):
 class SetCustomEnvironmentVariablesRequest(JsonResponse, APIRequest):
     method = "POST"
     url = "/api/v1/website/{website_id}/env/{stage}/environment-variables/custom/"
+
+
+# Repository
+
+
+class RepositoryRequest(JsonResponse, APIRequest):
+    url = "/api/v2/repositories/?website={website_id}"