+++ divio_cli/upload/boilerplate.py
@@ -10,6 +10,7 @@ from ..validators.boilerplate import validate_boilerplate
 from ..validators.common import load_config
 from .common import add_meta_files
 
+
 BOILERPLATE_EXCLUDE_DEFAULTS = ["boilerplate.json", ".git"]