+++ CHANGELOG.rst
@@ -1,20 +1,30 @@
 Changelog
 =========
 
+3.4.0 (2019-02-12)
+------------------
+
+ * Added remote git repository support
+ * Added testing
+ * Fixed issue on error handling while pulling files
+ * Removed binary builds
+ * Removed "cheatsheet" command
+
+
 3.3.12 (2019-01-09)
-------------------------
+-------------------
 
 * Updated DNS check to be backwards compatible
 
 
 3.3.11 (2019-01-09)
-------------------------
+-------------------
 
 * Updated DNS check to work with the latest busybox image. Older busybox versions must upgrade!
 
 
 3.3.10 (2019-01-07)
-------------------------
+-------------------
 
 * Fixed windows build
 
@@ -67,7 +77,7 @@ Changelog
 
 
 3.3.1 (2017-07-06)
------------------
+------------------
 
 * Minor bug fixes and automation improvements